Integrated oxygen alert into idle notifier
